REFINE: rename value object modules for consistency and implement new value objects for ISBN, rating, review text, tag name, and user email
