+++ src/domain/book/value_objects/book_isbn.py
